move `MaterialComposite`; add `fun` attribute for lab-plot
from _material.py to _template.py
