+++ matadi/__init__.py
@@ -6,9 +6,12 @@ from ._material import (
     Material,
     Material as MaterialScalar,
     MaterialTensor,
-    MaterialComposite,
 )
 from . import models
 from . import math
-from ._templates import ThreeFieldVariation, MaterialHyperelastic
+from ._templates import (
+    ThreeFieldVariation, 
+    MaterialHyperelastic, 
+    MaterialComposite,
+)
 from ._lab import Lab
